test: don't test that we can reproduce a bug under non-default settings
This removes a test case that tries to observe that something doesn't
happen when using non-default settings. The test was already broken
because it did not change the setting of `enableAsyncTasks` to false.
Instead of trying to fix the test so that it can configure the stream
processor and thus observe this behavior, I've chosen to simply remove
the test as it does not add any value.