refactor: arch-agnostic Dockerfile
Refactors the `Dockerfile` to be architecture/platform agnostic. Since
images have a general digests, separate from their platform-specific
digests, we can simply target that one. When building that image on the
specific platform (or even cross-building) the appropriate image will be
selected anyway, so we don't need to keep track of platform specific
digests.

One caveat is that it seems the ARM64 JDK image fails to build a custom
JRE without specifying a different process launching mechanism. It's a
bit unclear to me why, but apparently it's not an uncommon problem, and
the simplest option is to change the launch mechanism. Naja, not
perfect, but works.

(cherry picked from commit 82d1f59863089d8fc82c4fd9e69a5f2a3d396c48)