refactor(clients/java): use <default> tenant id by default
Ensure that the <default> tenant is used if no tenant id is
specified for a command. This value will be set through the
defaultTenantId Zeebe client configuration properties. The
value will be used always, i.e. when multi-tenancy is enabled
or disabled as we will map data to the <default> tenant if
nothing is configured.

Furthermore, this commit ensures that deprecated constructors
will map data to the <default> tenant if they are used in the
future.