feat(gateway): expand tenant id validation
When multi-tenancy is disabled, allow for users to submit
either an empty string, or the <default> tenant id on
their client request. The goal is to make users aware
that when multi-tenancy is disabled, all of their data
is owned by the <default> tenant, even though they don't
use the feature.

This commit also adds an explicti dependency to Apache
commons-lang3 since we use the library to validate if
a tenantId property has any text in it's value.