ci(.github): run daily-qa only on weekdays
We see several reasons to run the daily QA only on weekdays:

When the QA fails, it notifies the medic through slack. This means that
the medic can also be notified in the weekend. This appears to happen
often and the team experiences this as annoying.

It is also a lot of additional work for the medic when they start their
shift on Monday. Immediately being confronted with a bunch of QA run
failures is demotivating. Finding them during the week would be fast
enough, and avoids starting the week with an immediate downer.

Running the QA brings costs, running it less often reduces these costs
and reduces the carbon footprint.

In the weekend, it is not expected for the code under test to change. So
weekend runs don't add much over weekday runs.

Running them only on weekdays reduces the number of chances these tests
have to find bugs from 7 times per week to 5. The team thinks this is
still enough chance to find bugs.

(cherry picked from commit 1289af40e693c0303106be71e33135277885f7bb)