feat(protocol-impl): add tenantId property to pi-related record values
Implement the tenantId property to the ProcessInstanceCreationRecord
and the ProcessInstanceResultRecord. This property declares the
tenant ownership of the process instance that will be created for the
former, or the process instance that was created for the latter record.

The default value for the tenantId property is set to <default> as
this is the identifier of the tenant that will own  any Zeebe data
by default when multi-tenancy is disabled.