refactor: directly write to the snapshot dir
Instead of writing to a pending directory we directly write to the snapshot directory. This allows us to remove the move
and we already make sure that a snapshot is only valid because of an existing checksum.

If we receive multiple snapshots with same id we delete the previous one, if it is not persisted yet.