feat: introduce MigrationTransitionStep
Pull out migration of engine state from StreamPlatform to
PartitionTransition.

Previously, migration was done only if StreamProcessor was installed
with processing in mind (leader only).

The new migration step will allow to migrate the state in all cases and
role changes.

When transitioning from Follower to Leader, the database is reused,
which means migration has happened already and is automatically skipped.

If follower receives a snapshot, partition goes to INACTIVE and FOLLOWER
again and this approach also allows to migrate received old snapshots.

Only if Leader becomes follower, and state is recreated from snapshot
migration has to run again. Or on broker restart.

Migration can take time, and is blocking the partition installation.
This is the reason why we need to decouple the ready status from the
Zeebe partition installation.