refactor: wrap tenant id key in order
The tenant aware keys in the process state all have the tenant id
prefixing the respective key. Thus it makes sense to always wrap the
tenantId first.

This does not change any behavior, but this order can be expected by the
reader. This same order was already applied to ProcessVersionManager.