refactor: pass multiTenancyCfg to request mapper
The request mapper needs to know whether multi-tenancy is enabled or
not. However, the methods in the request mapper are all static, and the
EndpointManager expects them to be mapper functions (from A to B). This
makes it difficult to pass in the boolean flag whether multi-tenancy is
enabled or not.

Since this configuration does not change at runtime (it's read once at
startup) we can set it statically during the creation of the
EndpointManager. This is threadsafe, because the EndpointManager is the
only production code using the RequestMapper and the RequestMapper's
methods are only called after the EndpointManager is fully constructed.

We've additionally documented this usage restriction in JavaDoc as we
explicitly don't want this to change. So it's expected that the comment
will not go out-of-sync. Alternatively, an ArchUnit test could be added
for this, but we felt this goes too far. The JavaDoc is clear enough.

Alternatively, we could extract the mapping validations
(ensureTenantIdSet and ensureJsonSet) out of the RequestMapper. We see
this is a potential future improvement. There is not yet a direct need
to change this now.

The StubbedGateway uses a default MultiTenancyCfg for now, that will
change when we improve the GatewayTest.