test: allow configuring the GatewayTest
The gateway test was a bit hard to configure because calling super() in
the constructor does not allow for much modification of the GatewayCfg.

Instead, we can use a GatewayCfg Consumer as parameter in the
constructor. This allows for patters like:

```java
MyGatewayTest() {
  super(cfg -> cfg.setSomeConfig(true))
}
```

This is useful when we start testing multi-tenancy related topics in the
gateway and we need to selective enable or disable the multi-tenancy
feature flag