test: use IdentityInterceptor in GatewayTest
The GatewayTest sets up its own Gateway to run the tests against. This
gateway did not yet have the IdentityInterceptor setup. So none of the
tests (except for the IdentityInterceptorTest which are unit tests)
actually passed through the IdentityInterceptor.

To be able to verify that the authorized tenants are set correctly on
the broker request, we'll need to be able to control the Identity sdk
used by the IdentityInterceptor as well.

This commit adds the IdentityInterceptor to the StubbedGateway, and
adds control for the mocked Identity used by the IdentityInterceptor.