test: set fake bearer token on gateway test requests
Now that the IdentityInterceptor is used on all requests passing the
StubbedGateway in GatewayTests, the requests are all rejected because
they are missing the Authorization header specified by the
OAuthCredentialsProvider.

This header should contain a bearer token which the IdentityInterceptor
validates. When missing, or Identity considers it invalid, then the
request is rejected.

The FakeOAuthCallCredentials make sure a fake bearer token is set on
each client call. You can control the actual token using the setter, but
for now this is unused. Most tests will be fine with the "Bearer token"
value.

Note that Identity is mocked and can be controlled from the
StubbedGateway.