docs: add zeebe.gateway.multiTenancy.enabled config
This adds a description for the config option to enable multi tenancy.

Note that currently only the Identity tenant authorization provider is
supported. As such there is no way to specifically select the
authorization provider. When we add support for custom tenant
authorization providers through custom interceptors, we need to add a
config option to the multiTenancy section to select it. Otherwise, it
works magical, i.e. the user might want to use Identity for
authentication but not as authorized tenant provider.