test: add missing env var
Seems that the long polling enabled was not overwritten by an env var in
the test case. This did not surface in the test case as a failure,
because the used config (customConfig) also used false as value.

By overwriting the value to true and expecting this value, we actually
verify that the env var overwrites the configuration defined in the
file.

The expected order is: Env Var > Config File > Default