feat: mark broker ready after all partitions joined raft
Introduce new PartitionRaftListener, which notifies when the parition succesfully changed a role in raft.
This can be follower or leader. This means the listener is notified earlier as the existing PartitionListeners,
which are notified AFTER the zeebe partition is completely installed.

This allows to mark the broker earlier ready, which helps in rolling updates, especially if replay or migration
takes longer.