feat(qa): add test broker launcher
Adds a test application launcher for the `StandaloneBroker`, which
allows configuring the `BrokerCfg` bean programmatically, probing the
health actuator, and awaiting it's various states (e.g. topology
complete, broker ready, etc.), as well as common helpers to add
exporters, the recording exporter, etc.

Additionally, a new interface is added to identify apps that can act as
gateways (i.e. the broker and standalone gateway). This will become
particularly useful when we start with the cluster API.