test: encode default tenant when disabled
When multi-tenancy is disabled, the gateway should encode the default
tenant as the authorized tenant on the JWT.

I've taken the opportunity to rename the test case that verified what
should happen when multi-tenancy is enabled.

First, this is guided by the test infra which does no easily allow
disabling multi-tenancy in between test cases of the same test class.

Second, we can use these test classes to fit in most of the general
integration tests for multi-tenancy in the gateway. For example, that
specific requests require setting the tenant id when multi-tenancy is
enabled, or that they disallow it when it is disabled. Another example
is that these could host the tests where the request is rejected because
the specified tenant is not one of the authorized ones. We can add such
cases later.