feat(qa): add test cluster utility
Adds a new utility to build and launch clusters of `TestApplication`,
e.g. `TestStandaloneBroker`, `TestStandaloneGateway`, etc. Allows
configuring the cluster in coarse ways (e.g. setting the cluster name
across the name, the # of brokers, the replication factor), or fine
grained via `withBrokerConfig`, `withGatewayConfig`, and
`withNodeConfig`.

Minimal tests were added to the builder, as there it's useful to know if
the logic of building your clusters is broken since it won't be
immediately obvious by using the `TestCluster` that such things broke.