refactor: fail early
The most complicated code in this method happens when multi-tenancy is
enabled. We can first deal with the case where multi-tenancy is
disabled to get this out of the way. This makes the code easier to
understand, and we can remove the redundant else as well in another
commit. (In another commit to keep these commits readable).