test: await actor lifecycle started
Await the actor lifecycle to be completely started before we submit
jobs on different actors, otherwise the ordering is not guaranteed.

When running more then one actor, even with one thread, it might
happen that jobs (even submitted after each other) might run in a
different ordering.

Submitting an actor to the scheduler will actually submit in the
background the ActorTask to the WorkStealingGroup and notify the
ActorThread.

The Actors have a specific lifecycle which they run before they
actually executing any jobs which are submitted from outside.
If they complete that lifecycle they will end in the STARTED state
and put themself to waiting. If this is the case and a new jobs is
submitted, this means the corresponding ActorTask is resubmitted to
the WorkStealingGroup and the ActorThread is notified.

If the Actor is still running like running the last job in the
lifcecyle to start the Actor and the new job was already submitted
it will directly consume it via polling the submitted queue.

Imagine the following ordering:

Task 1 - Actor submit
Task 2 - Actor submit

T1 - onActorStarting
T2 - onActorStarting

T1 - STARTED <- Try Wait
   | When now a job is submitted here the task is resubmitted to
   | the queue to run the job
T2 - STARTED
   | When a job for T2 was submitted before it went to the
   | waiting state it will immediately be consumed before T1 can run

T1 - job