fix: reduce scope for AppendListener
In order to remove a memory leak we need to reduce the references to the indexed record.
This commit adjustes the AppendListener#onCommit to only pass in the index instead of the whole object.

(cherry picked from commit d57c39d2b8a9bb09e1484c9e55b25ddd66835d12)