fix: use permission denied for unauthorized tenant
When the specified tenant doesn't match any of the authorized tenants,
then the request should be rejected with the PERMISSION_DENIED status
code. In all other cases, we should reject with the INVALID_ARGUMENT
status code.

This also added a new GatewayAssertions test helper class, that we can
use to store such assertions.