refactor: rework RaftCommittedEntryListener
Iterate over RaftCommittedEntryListener and refactor the listener such it serves the actual need.

We have some services (to be specific the AsyncSnapshotDirector) which need the committed position, and
want to listen to new updates. In raft we know which record we are committing and whether it was an application record so we can pass this information threw the listeners.

This avoids to pass in the whole IndexedRecord object, and reduce the potential of going out of OOM because of keeping to much data in heap (when commit is not possible).

(cherry picked from commit daad65c045b120f91490fadcf00278afa0a4ebfa)