fix: add auth info to all sendRequest methods
Not all send request methods in the EndpointManager were encoding a JWT
for the authorization property. For example, sending a create process
instance request would not have the authorization set on the broker
request.

By extracting this code, we can easily reuse in the other two places
where requests can be sent in the EndpointManager.

Alternatively, we could have re-abstracted the brokerClient and
sendRequestWithRetry fields to use a common interface, but this seemed
to much effort for now.