fix(engine): ensure tenant-aware timers are rescheduled correctly
On process definition deployment, timers are re-scheduled on the latest
version of the process definition. In order to do that, existing timers
check the id of the process definition they are part of with the newly
deployed process definition.

When multi-tenancy was introduced, it required to store the tenant id
of the process definition in the state. This impacts the lookup that
timers perform on re-scheduling since they also need to provide the
id of the tenant they belong to.

This commit ensures that the TimerInstance provides this information.