feat(gateway): start GatewayClusterTopologyService with BrokerClient
Periodic sync is disabled in the gossiper to avoid unnecessary noise
in the logs if the feature is disabled in the Broker. Even if sync is
disabled, the gateway can still receive gossip and sync request from
other Brokers. So it will eventually receive the updated ClusterTopology.

When the feature is permanently enabled in the Broker, we can also enabled
sync here.