fix: don't send requests to closed members
Response handlers for append requests hold onto a reference to the
member context, even if that member context was closed in the meantime,
for example because the member was removed from the configuration.

To prevent making new requests for a member not in the current
configuration, we mark the context as closed. While it is closed, the
member cannot receive append, configure, or install requests.
Opening a replication context for a closed member fails too.