fix: only try to join via members in the replication group
A joining member needs a set of members that can handle a join request.
We used to use the membership service to just get all members of the
cluster and try all of them.
This was not correct because the membership service does not know about
partition distribution and the joining member would end up sending join
requests to members which do not even listen to requests for the
partition.
Instead, we now use the partition metadata to find members that can
handle the join request.
In tests we can provide a static list of known members.