test: support multi tenancy over identity
This adds extensive E2E tests that verify that multi-tenancy behaves as
expected in combination with Identity as the Authorized Tenants
Provider.

These tests spin up 4 containers:
- Zeebe Broker with Embedded Gateway, configured to connect to Identity
- Identity configured to work with Keycloak and storing data in a
  POSTGRES database
- Keycloak configured to store date in a POSTGRES datebase
- POSTGRES database for Keycloak and Identity

Each of the test cases use specific ZeebeClients that use Identity for
authentication. They are associated within Identity to specific tenants.
This means that the token in the request received by the Zeebe Gateway
is one that Identity has provided and understands. I.e. Identity can
tell the Gateway which authorized tenants are associated to the token.

In the test cases the specific ZeeebeClient to use can be defined by
passing a specific clientId. See `createZeebeClient`. The used names are
supposed to clarify which tenants are associated to the specific client.

To set all this up, several environment variables are defined for
Identity, in the form of `KEYCLOAK_CLIENTS_*`. These make sure API
clients are registered within Keycloak. Before the tests run, each of
these clients is associated to specific tenants (see `init()`). This
requires some POSTGRES SQL statements, as the associations cannot be
configured through environment variables **YET**.