fix(broker): dont expect authorization on request
Not all Execute Command Requests will have the authorization property
specified. For example, when an older gateway or broker is communicating
with an upgraded broker.

In tests this also happens often, so we can't just check the sbe schema
version. However, we can check whether the buffer contains more bytes
after the value, or not. If it does, then this must be the
authorization. This is guaranteed because SBE requires ordered encoding
and decoding of variable length data.

>Note: Variable data fields must be encoded and decoded in order as
>defined in the schema.

See: https://github.com/real-logic/simple-binary-encoding/wiki/Java-Users-Guide#variable-length-data