test: unflake MultiTenancyOverIdentityIT
The tests weren't waiting for Identity to initialize the clients in
postgres. While the tests worked on my machine, they didn't on the CI
because Identity had not yet completed initialization.

We can simply throw an exception when any of the postgres queries
doesn't provide results. We can deal with the exception using
Awaitility, retrying our own initialization (associating the tenants
with clients) until no exception is thrown.

Retrying every second seems reasonable for an IT, and we can fail if it
didn't succeed after an entire minute of retrying.