test: adjust size of large process
The large process model that fits just within the max message size had
become larger than the max message size because the authorization
property of the record metadata is now filled with an actual value.

I was able to lower the value by 200 bytes, but the test would also have
worked at 1046849 (or 51 bytes reduced).

The additional 149 bytes gives some buffer when we add more metadata to
records, before this test fails again.