refactor(atomix): remove outdated test
Removes an outdated test which was incorrectly testing behavior that
does not exist anymore. We used to persist snapshots before resetting
the log, and this is what the test was covering. However, it never
worked because it was in fact restarting the node with no snapshot
(which is also why it didn't fail when we removed the old behavior).

As we now always persist snapshots after resetting the log, this case
simply cannot happen.