refactor(gateway): calculate partition count from ClusterTopology
Although we don't support changing number of partitions yet, it makes sense
to calculate it from the ClusterTopology as it is the source of truth. This
also helps to write test where we want to incorrectly configure some brokers.

For now, we use the static replicationFactor. During configuration change
there will be short period where number of replicas differ from the configured
one. So it is difficult to calculate what is the exact replicationFactor based
on current ClusterTopology.