test: Remove gateway incident step from generator
The process generator can generate processes with exclusive gateways that raise an incident on activating the gateway
because of a non-existing variable.

With the new version of the FEEL engine, a non-existing variable results in `null`. Previously, the evaluation failed
with an error.

Since it is not easy anymore to raise an incident on the gateway, we remove the step from the generator.