feat(engine): make job activation tenant-aware
Ensure that activatable jobs kept in the state are tenant-aware.
When querying for activatable jobs, the tenant ownership of the
job is considered as well.

To avoid a large migration of the JobState, a workaround is
implemented to map old (pre-tenant-aware) data to the <default>
tenant. This is done by storing, querying, and deleting this type
of data without a tenant id.