feat: allow adding more fields to broker request
If we ever want to expand the ExecuteCommandRequest with more
variable-length properties in the future, then this code would not allow
it.

Instead, we tried to use the example from `RecordMetadata.wrap`, but
that didn't work. Somehow we can't call `decoder.authorizationLength()`
when the authorization is unset.

What we can do however, is simply check if there is more to come in the
buffer, or whether the decoder has reached the end.