test: should not call process of other tenant
This was a missing test case.

As the user has no way to discover the incident, I've used the recording
exporter to assert the incident. An alternative approach would be to
set up the process so that the client could interact somehow with the
parent process after calling the process (e.g. a service task after the
call activity, with a pass through child process). However, I find that
much less clear. Additionally, as we start to test other concepts like
messages and such, we'll need to verify the exported log more often.
Lastly, this is totally fine, as the exported log is also part of the
API that users can use.