test(engine): verify version cache doesn't reference same objects
When we get an object from the persisted state it is always stored in a
the same object. This object is cached directly. As a result different
cache entries reference the same object. This causes different process
versions to influence each other.

This test verifies that if we have 2 processes and we clear the cache so
 we get them from the persisted state, the versioning still works as
 expected.