test: reproduce detected inconsistency
This test is the same as the shouldResolveIncident() test above it, but
adds a message boundary event to the call activity.

When activating the call the activity, we subscribe to the message
boundary event, but then raise an incident due to the child process not
being found.

However, the subscriptions are not rolled back. So when we resolve the
incident after deploying the child process, we attempt again to
subscribe to the message boundary event. This now leads to an
inconsistency check failure, as the message subscription already exists.

(cherry picked from commit cdd1a0c51f4c22d9737b8616cf7ecc4ee97b3f6e)