fix(engine): fix process versioning cache
The cache would directly store the value returned from the persisted
state. This value is always store in the same object. As a result
different processes in the cache could reference the same value and
versions could start influencing each other.

By making sure we copy the value before storing it in the state we can
gaurantee that each process gets it's own object in the cache.