test: detect inconsistency on resolve incident
Just like the test in the previous commit, the inconsistency can also
occur on any job worker task where any of the job worker expressions
fails to evaluate.

When it fails to evalute, it raises an incident, but only after
subscribing to any of the boundary events.

On resolving the incident, we subscribe again to the boundary events,
causing inconsistency failures as these subscriptions already exist.