fix: only subscribe to events if expr-eval succeeds
This resolves a critical bug for any job worker task, where the task is
subscribed to its events a second time when resolving any incident
created after subscribing to those events.

For example, a service task with a message boundary event may subscribe
to that event and then still fail the evaluation of the job type
expression. On resolving the incident, the activation of the service
task is continued which would result in subscribing the task to the
event again.

By first evaluating the job expressions, we can safely create the
incident without any subscriptions. Note that the call to
subscribeToEvents itself first evaluates all event expressions, before
creating the subscriptions. So that part is safe as well.

This is only a patch to cover up a problem. Eventually, we need a safer
way of creating incidents without any other effects.

(cherry picked from commit c346cae8bcda7455ab6a624d1e9cea8704f751f8)