test: speed up slow test
This test was waiting for the process instance to complete, which was
taking up to a second on my machine. But we don't have to await that.
All we need is that the element is activated. That already indicates
that the incident was resolved and no new one was created. This speeds
up the test to run within 50ms.