test: speed up slow test
This test was waiting for the process instance to complete, which was
taking up to a second on my machine. But we don't have to await that.
All we need is that the element is activated. That already indicates
that the incident was resolved and no new one was created. This speeds
up the test to run within 50ms.

(cherry picked from commit 0e9c3b933779f5a4b093b6950fb7872c28ab6265)
(cherry picked from commit 3bf781468d2cc4e37900967f472f3055cc59b140)