feat: Simplify variable output mapping
The behavior changed in the new version of the FEEL engine. A non-existing variable doesn't fail the evaluation anymore.

The only way to fail the evaluation is to use the new `assert()` function. The usage of the `is defined()` function is
not required anymore because the function works in all cases.

Update the deprecated function `put all()` by the new function `context merge()`.