fix(engine): set correct tenant id on job state operations
Due to the workaround to not set the JobState tenant id
when data is owned by the <default> tenant, an old value
of the tenant id is used, which leads to bugs in the
job activation behavior. To fix this, we must explicitly
set the value of the tenant id to an empty string.