test: copy legacy process state
This class can be used in tests for migrating the process state from 8.2
to 8.3.

It's mostly a direct copy, with two exceptions:
- I've inlined the ProcessVersionManager
- I've referred to the DEPRECATED column families

While not all code in this class may get used, it felt much easier than
writing a new legacy state class that has to store the state exactly as
this copy.