refactor: extract DbProcessMigrationState
The DbMigrationState is growing fast, and will grow out of control if we
don't take action.

By extracting a specific migration state class (the one for
DbProcessState to 8.3), we can keep each state migration somewhat under
control. At a maximum, the class contains twice as many column families
as the corresponding state class under migration (once for the legacy
column family and once for the new column family).