refactor: make legacy state available for prod
Instead of having to add each individual column family to the
DbMigrationState or a new DbDecisionMigrationState, it makes much more
sense to use a copy of the LegacyDecisionState to access the column
families.

I worried that I would make small mistakes in the column family
definitions, and that problem is now completely gone.