feat(engine): provide <default> tenant id for job activation
Provide the <default> tenant id for job activation by job polling when
no tenant ids have been provided. The Gateway prevents the client from
providing an empty list of tenant ids for activation, but it's still
possible if an old Gateway is used that doesn't support multi-tenancy.
In this case, we should provide the <default> tenant id in the engine,
in the JobBatchCollector, as this is where these ids are relevant.