test: ensure decision state migrated
This verifies that each of the 6 column families of DbDecisionState are
migrated, by:
- setting up data in the corresponding LegacyDecisionState's column
  family
- running the migration tasks
- using the DbDecisionState methods to verify the current state

I had considered exposing the DbDecisionState's column families directly
for the tests, but I found this too time intensive. For pragmatism, I
reverted to using the available methods.

Note that it was difficult to assert the
decisionKeyByDecisionIdAndVersion column family, as there are no
available find/get/visit methods for it. Instead, it is only used
internally when deleting the latest decision, so that we can find the
key of a previous decision. It then becomes the latest decision. I've
used this insight knowledge in the test, which is less than great. But,
it was the most pragmatic solution that I could think of.