feat: migrate decision state for multitenancy
This introduces a migration of decision state for multitenancy.

Originally, I wanted to avoid re-defining all target column families by
exposing the DbDecisionState's column families. However, that meant I
had to also expose all the relevant keys (to actually use these column
families). This became unwieldly, unless I grouped the keys that belong
to a specific column family. But this became time consuming.

A second (potentially better) approach would be to refactor the
DbDecisionState to simplified methods (insert into column family x, find
from column family y, etc.). We've previously discussed this approach to
help avoid making changes in how events are applied, because the changes
are then required to be made in an event applier. That would then
require a new event applier. See #13813 for more details. That approach
would allow direct access of each the column families of the state
class.

However, since I don't want to spend too much time on this, I've opted
for the surprisingly faster solution of copy-pasting the existing keys
and column families.

The migrations themselves are straightforward, although sometimes hard
to read due to the usage of foreignkeys that wrap the keys used in each
of column family visitors.