refactor(clients/java): adjust default job worker tenant ids
The defualtJobWorkerTenantIds configuration property by default
contains a singleton list with the <default> tenant id. This will
be used when multi-tenancy is disabled or if users keep using the
<default> tenant when multi-tenancy is enabled.

The defaultJobWorkerTenantIds configuration property uses a comma
(,) as the list delimiter, instead of a comma and space (, ). We
decided on this because it would be more difficult to define a list
with spaces in the command line.