test: give all clients access to default tenant
Typically, we expect that Identity gives each client access to the
default tenant. This was not modeled in this IT.

While it's unlikely that we should encounter that a client has NO
tenants associated (not even the default), it makes sense to keep
ZEEBE_CLIENT_ID_WITHOUT_TENANT to mean specifically without even the
default tenant. Otherwise, the currently disabled test case
shouldDenyDeployProcessWhenNoTenantAssociated would pass.

In practice, this doesn't affect any of the setup test cases, because
there were no tests dealing with the default tenant yet.