fix(raft): reset term from last log if metastore is empty
After restoring from a backup, metastore will be empty because backup
contains only journal segments. Since metastore is empty, previously raft
node re-starts the term at 0. This caused issues in election, because
accepting or rejecting a poll request is also based on the last log entry's
term. When the term restarts at 0, the last log entry's term can be higher than
the local term which results in unexpected election results. To prevent this
case we have to start the term at the term of the last log entry.

Alternate is to include meta file in the backup. But this doesn't have
much benefit because no other data in the metastore is useful after restore.
In some cases, it might even cause issues if restore starts with a different
configuration than in the backup.

(cherry picked from commit fb4bd88f85f24f2be33a62838ade11bb9e623b26)