feat(engine): adjust LegacyMessageSubscriptionState for usage
The class was a direct copy. It should not implement any interface, as
this has been made tenant-aware. It should also make use of the legacy
ColumnFamily, which it exposed via a getter so we can use it during the
migration.

All methods except for the put has been removed from the state. We need
to keep the put as we'll use it to insert data during our migration
tests.