fix(engine): use correct tenant
We were trying to find the latest form by id for the tenant in the
metadata record. This metadata record was only just created. No tenant
was set on it, so it would always result in the default tenant.

We must take the tenant from the deployment record. This is the tenant
the user provided when doing the deployment.