test(qa): align image version with sdk version
The snapshot image is not compatible with the e2e tests anymore. This
commit aligns the image version with the version we have defined in the
pom file.

We should improve this, getting the version of the sdk at runtime so we
don't have to manage the version in these tests. I have prepared a PR
for the identity sdk to make this possible:
https://github.com/camunda-cloud/identity/pull/2148