fix: migrate procDef based on parentChildCF
Previously, we ran the
migrateElementInstancePopulateProcessInstanceByDefinitionKey only if the
CF is empty and the target isn't. However, there's a possibility that
both are not empty while not all entries have been migrated.

Instead, we can assert that all process instances have an entry in the
processInstanceKeyByProcessDefinitionKeyColumnFamily. We can verify this
by counting all entries in that CF, and comparing it to the count of all
process instances in the parentChildColumnFamily (those prefixed by -1).

Doing so, allows us to speed up the migration task as well, by iterating
only over the process instances in the parentChildColumnFamily. Note
that this requires an additional lookup of the processInstance to find
the processDefinitionKey. While that may be slower, the entire migration
task is likely faster, as we don't have to iterate over all element
instances anymore.