fix: run migrations in transaction
The migration transition step was not run inside a transaction. This
caused problems on replay in combination with migrations, where the
changes from the migration were not available during replay. Whether the
transaction had been rolledback, or were simply not yet available to the
transaction at replay is unclear to me.

But we can safely start a transaction during the migration step and
commit or roll it back as needed.

I've tested this successfully manually. I don't see a chance to easily
test this in an automated fashion, so I'd rather not add any cases now
to make progress and unblock the release.